Replace otlp-proto-common encoding with the pure-Python protobuf backend
Move the shared OTLP encoders into a private
opentelemetry.exporter.otlp._proto.common package and keep the public
opentelemetry.exporter.otlp.proto.common.* API as re-export shims over it.
This mirrors the layout opentelemetry-proto already uses, where
opentelemetry._proto holds the implementation and opentelemetry.proto
re-exports it.

hatch now reads the version from the private package, matching
opentelemetry-proto.

The equivalence tests and the encoder benchmark are dropped. They imported
the public path expecting a protobuf-backed encoder and the private path for
the pure-Python one, which only worked while those lived in two separate
distributions. Now that one distribution owns both paths, the public path is
a shim over the private one, so every such comparison would compare the
pure-Python encoder against itself and pass without testing anything.
