+++ exporter/opentelemetry-exporter-otlp-proto-common/pyproject.toml
@@ -34,7 +34,7 @@ Homepage = "https://github.com/open-telemetry/opentelemetry-python/tree/main/exp
 Repository = "https://github.com/open-telemetry/opentelemetry-python"
 
 [tool.hatch.version]
-path = "src/opentelemetry/exporter/otlp/proto/common/version/__init__.py"
+path = "src/opentelemetry/exporter/otlp/_proto/common/version/__init__.py"
 
 [tool.hatch.build.targets.sdist]
 include = [
