+++ exporter/opentelemetry-exporter-otlp-proto-common/src/opentelemetry/exporter/otlp/proto/common/_log_encoder.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from opentelemetry.exporter.otlp.proto.common._internal._log_encoder import (
+from opentelemetry.exporter.otlp._proto.common._internal._log_encoder import (
     encode_logs,
 )
