+++ exporter/opentelemetry-exporter-otlp-proto-common/src/opentelemetry/exporter/otlp/proto/common/metrics_encoder.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from opentelemetry.exporter.otlp.proto.common._internal.metrics_encoder import (
+from opentelemetry.exporter.otlp._proto.common._internal.metrics_encoder import (
     encode_metrics,
 )
