+++ exporter/opentelemetry-exporter-otlp-proto-common/src/opentelemetry/exporter/otlp/proto/common/version/__init__.py
@@ -1,4 +1,7 @@
 # Copyright The OpenTelemetry Authors
 # SPDX-License-Identifier: Apache-2.0
 
-__version__ = "1.45.0.dev"
+
+from opentelemetry.exporter.otlp._proto.common.version import __version__
+
+__all__ = ["__version__"]
